Refactor view out of AreaView and StyleService out of GameService
