+++ src/IncrementalSociety.Web/Pages/AreaBuilding.razor
@@ -1,21 +1,18 @@
+@using System.Collections.Immutable
+
 @using IncrementalSociety.Web.Services
 @using IncrementalSociety.Model
 
+@inject GameService GameService
+@inject StyleService StyleService
+
 <div>
-	<div class="d-flex flex-column area-building" style="background-color:@BackgroundColor;" onclick=@OnClick>
-		<p class="area-building-title" >@Building</p>
-		@if (HasBuilding) {
-			<div class="d-flex flex-wrap">
-				@foreach (var res in GameService.GetBuildingResources (@Building))
-				{
-					<div class="building-resource d-flex align-items-center">
-						<img class="resource-image" src="@GameService.GetImageFilename(@res.Key)" />
-						<div class="@GameService.GetResourceDeltaClass(res.Value)">@res.Value</div>
-					</div>
-				}
-			</div>
-		}
-	</div>
+	<Building 
+		BuildingName="@Building"
+		Resources="@Resources"
+		BackgroundColor="@BackgroundColor"
+		OnClick="OnClick"
+		ColorResourcesAsDelta=true />
 </div>
 
 @functions {
@@ -28,10 +25,9 @@
 	[Parameter]
 	string Building { get; set; }
 	
-	[Parameter]
-	GameService GameService { get; set; }
-
 	bool HasBuilding => !String.IsNullOrEmpty (Building);
+
+	ImmutableDictionary<string, double> Resources => HasBuilding ? GameService.GetBuildingResources (Building) : null;
 
 	string BackgroundColor
 	{
