+++ src/IncrementalSociety.Web/wwwroot/css/site.sass
@@ -59,17 +59,6 @@
 	margin-left: 5px
 	margin-right: 5px
 
-.area-building
-	border: 1px solid black
-	padding-left: 5px
-	padding-right: 5px
-	margin-bottom: 5px
-	min-height: 35px
-	max-width: 200px
-
-.area-building-title
-	margin-bottom: 5px
-
 #resource-list
 	list-style-type: none
 	padding-left: 0px
@@ -106,6 +95,17 @@
 
 .clear
 	display: none
+
+.building
+	border: 1px solid black
+	padding-left: 5px
+	padding-right: 5px
+	margin-bottom: 5px
+	min-height: 35px
+	max-width: 200px
+
+.building-title
+	margin-bottom: 5px
 
 .building-resource
 	font-size: .75rem
